Se mueve la lógica getCases a un custom hook
